# openapi.json
{
  "components": {
    "schemas": {
      "AssetOut": {
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "exchange": {
            "title": "Exchange",
            "type": "string"
          },
          "interval": {
            "title": "Interval",
            "type": "string"
          }
        },
        "required": [
          "asset",
          "exchange",
          "interval"
        ],
        "title": "AssetOut",
        "type": "object"
      },
      "BacktestRequest": {
        "description": "Re-run a saved strategy on the run's original windows.\n\n``train_window``/``validation_window``/``test_window`` are intentionally\nNOT exposed yet \u2014 Phase 5 will add cross-asset / cross-window\nre-evaluation, with whitelisted asset registration. For now, the only\nsupported re-run is on the same windows the strategy came from.",
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "strategy_id": {
            "title": "Strategy Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "strategy_id"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BacktestResponse": {
        "properties": {
          "buy_and_hold_test": {
            "$ref": "#/components/schemas/PerformanceMetricsOut"
          },
          "chosen_strategy_id": {
            "title": "Chosen Strategy Id",
            "type": "string"
          },
          "overfitting_gap": {
            "title": "Overfitting Gap",
            "type": "number"
          },
          "random_entry_test": {
            "$ref": "#/components/schemas/PerformanceMetricsOut"
          },
          "test_metrics": {
            "$ref": "#/components/schemas/PerformanceMetricsOut"
          },
          "train_metrics": {
            "$ref": "#/components/schemas/PerformanceMetricsOut"
          },
          "validation_metrics": {
            "$ref": "#/components/schemas/PerformanceMetricsOut"
          }
        },
        "required": [
          "chosen_strategy_id",
          "train_metrics",
          "validation_metrics",
          "test_metrics",
          "buy_and_hold_test",
          "random_entry_test",
          "overfitting_gap"
        ],
        "title": "BacktestResponse",
        "type": "object"
      },
      "CreateRunRequest": {
        "description": "Body of POST /runs.\n\n``asset`` must be a registered asset (see assets.py). ``population_size``\nis also the number of strategies the server generates.",
        "properties": {
          "asset": {
            "title": "Asset",
            "type": "string"
          },
          "elitism_count": {
            "default": 2,
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Elitism Count",
            "type": "integer"
          },
          "generations": {
            "default": 50,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Generations",
            "type": "integer"
          },
          "population_size": {
            "default": 10,
            "maximum": 200.0,
            "minimum": 2.0,
            "title": "Population Size",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "selection_pressure": {
            "default": "tournament",
            "title": "Selection Pressure",
            "type": "string"
          },
          "slippage_bps": {
            "default": 1.0,
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "type": "number"
          },
          "stop_loss_pct": {
            "default": 0.0075,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Stop Loss Pct",
            "type": "number"
          },
          "taker_fee_bps": {
            "default": 10.0,
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Taker Fee Bps",
            "type": "number"
          },
          "target_pct": {
            "default": 0.015,
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Target Pct",
            "type": "number"
          },
          "trade_window_bars": {
            "default": 96,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Trade Window Bars",
            "type": "integer"
          }
        },
        "required": [
          "asset"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
      "CreateRunResponse": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status",
          "status_url"
        ],
        "title": "CreateRunResponse",
        "type": "object"
      },
      "GenerationDetailOut": {
        "properties": {
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/StrategyOut"
            },
            "title": "Strategies",
            "type": "array"
          },
          "train_metrics": {
            "$ref": "#/components/schemas/GenerationMetricsOut"
          },
          "validation_metrics": {
            "$ref": "#/components/schemas/GenerationMetricsOut"
          }
        },
        "required": [
          "run_id",
          "number",
          "train_metrics",
          "validation_metrics",
          "strategies"
        ],
        "title": "GenerationDetailOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunDetailOut": {
        "properties": {
          "buy_and_hold_test": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "chosen_strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Strategy Id"
          },
          "current_generation": {
            "title": "Current Generation",
            "type": "integer"
          },
          "generations": {
            "items": {
              "$ref": "#/components/schemas/GenerationSummaryOut"
            },
            "title": "Generations",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manifest": {
            "$ref": "#/components/schemas/ManifestOut"
          },
          "overfitting_gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overfitting Gap"
          },
          "random_entry_test": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "test_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "train_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "validation_metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PerformanceMetricsOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "current_generation",
          "manifest",
          "generations",
          "chosen_strategy_id",
          "overfitting_gap",
          "train_metrics",
          "validation_metrics",
          "test_metrics",
          "buy_and_hold_test",
          "random_entry_test"
        ],
        "title": "RunDetailOut",
        "type": "object"
      },
      "RunSummaryOut": {
        "properties": {
          "chosen_strategy_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Strategy Id"
          },
          "current_generation": {
            "title": "Current Generation",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "overfitting_gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overfitting Gap"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "current_generation",
          "started_at",
          "finished_at",
          "seed",
          "chosen_strategy_id",
          "overfitting_gap"
        ],
        "title": "RunSummaryOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "gentrade",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/assets": {
      "get": {
        "operationId": "get_assets_assets_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetOut"
                  },
                  "title": "Response Get Assets Assets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Assets"
      }
    },
    "/backtests": {
      "post": {
        "operationId": "post_backtest_backtests_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Backtest"
      }
    },
...
    "/runs": {
      "get": {
        "operationId": "get_runs_runs_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummaryOut"
                  },
                  "title": "Response Get Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Runs"
      },
      "post": {
        "operationId": "post_run_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Run"
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/runs/{run_id}/generations/{n}": {
      "get": {
        "operationId": "get_generation_runs__run_id__generations__n__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "n",
            "required": true,
            "schema": {
              "title": "N",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Generation"
      }
    },
    "/runs/{run_id}/strategies/{strategy_id}": {
      "get": {
        "operationId": "get_strategy_runs__run_id__strategies__strategy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy"
      }
    }
  }
}